Fix: Cast string length from size_t to short to avoid C4267 warning

+++ Cpp-Library-Collection/MyString.h
@@ -34,7 +34,7 @@ public:
 
     static short Length(std::string S1)
     {
-        return S1.length();
+        return static_cast<short>(S1.length());
     };
 
     short Length()
